Move site configs config to new provider
