+++ demo/admin/src/App.tsx
@@ -13,7 +13,6 @@ import {
     LocaleProvider,
     MasterMenuRoutes,
     SitePreview,
-    SitesConfigProvider,
 } from "@comet/cms-admin";
 import { css, Global } from "@emotion/react";
 import { createApolloClient } from "@src/common/apollo/createApolloClient";
@@ -85,90 +84,87 @@ export function App() {
                     DamFile: createDamFileDependency(),
                 },
             }}
+            siteConfigs={{
+                configs: config.siteConfigs,
+                resolveSiteConfigForScope: (configs, scope) => {
+                    const siteConfig = configs.find((config) => {
+                        return config.scope.domain === scope.domain;
+                    });
+
+                    if (!siteConfig) throw new Error(`siteConfig not found for domain ${scope.domain}`);
+                    return {
+                        url: siteConfig.url,
+                        preloginEnabled: siteConfig.preloginEnabled || false,
+                        blockPreviewBaseUrl:
+                            siteConfig.scope.domain === "secondary"
+                                ? `${siteConfig.url}/block-preview`
+                                : `${siteConfig.url}/block-preview/${scope.domain}/${scope.language}`,
+                        sitePreviewApiUrl: `${siteConfig.url}/site-preview`,
+                    };
+                },
+            }}
         >
             <ApolloProvider client={apolloClient}>
                 <BuildInformationProvider value={{ date: config.buildDate, number: config.buildNumber, commitHash: config.commitSha }}>
-                    <SitesConfigProvider
-                        value={{
-                            configs: config.sitesConfig,
-                            resolveSiteConfigForScope: (configs, scope) => {
-                                const siteConfig = configs.find((config) => {
-                                    return config.scope.domain === scope.domain;
-                                });
-
-                                if (!siteConfig) throw new Error(`siteConfig not found for domain ${scope.domain}`);
-                                return {
-                                    url: siteConfig.url,
-                                    preloginEnabled: siteConfig.preloginEnabled || false,
-                                    blockPreviewBaseUrl:
-                                        siteConfig.scope.domain === "secondary"
-                                            ? `${siteConfig.url}/block-preview`
-                                            : `${siteConfig.url}/block-preview/${scope.domain}/${scope.language}`,
-                                    sitePreviewApiUrl: `${siteConfig.url}/site-preview`,
-                                };
-                            },
-                        }}
-                    >
-                        <IntlProvider locale="en" messages={getMessages()}>
-                            <LocaleProvider resolveLocaleForScope={(scope: ContentScope) => scope.language}>
-                                <BlocksConfigProvider
-                                    isBlockSupported={(block, scope) => {
-                                        if (scope.domain === "main") {
-                                            return true;
-                                        } else {
-                                            return (
-                                                block.name !== NewsDetailBlock.name &&
-                                                block.name !== NewsListBlock.name &&
-                                                block.name !== NewsLinkBlock.name
-                                            );
-                                        }
-                                    }}
-                                >
-                                    <MuiThemeProvider theme={theme}>
-                                        <DndProvider options={HTML5toTouch}>
-                                            <SnackbarProvider>
-                                                <CmsBlockContextProvider>
-                                                    <ErrorDialogHandler />
-                                                    <CurrentUserProvider>
-                                                        <RouterBrowserRouter>
-                                                            <GlobalStyle />
-                                                            <ContentScopeProvider>
-                                                                {({ match }) => (
-                                                                    <Switch>
-                                                                        <Route
-                                                                            path={`${match.path}/preview`}
-                                                                            render={(props) => (
-                                                                                <SitePreview
-                                                                                    resolvePath={(path: string, scope) => {
-                                                                                        return `/${scope.language}${path}`;
-                                                                                    }}
-                                                                                    {...props}
-                                                                                />
-                                                                            )}
-                                                                        />
-                                                                        <Route
-                                                                            render={() => (
-                                                                                <MasterLayout
-                                                                                    headerComponent={MasterHeader}
-                                                                                    menuComponent={AppMasterMenu}
-                                                                                >
-                                                                                    <MasterMenuRoutes menu={masterMenuData} />
-                                                                                </MasterLayout>
-                                                                            )}
-                                                                        />
-                                                                    </Switch>
-                                                                )}
-                                                            </ContentScopeProvider>
-                                                        </RouterBrowserRouter>
-                                                    </CurrentUserProvider>
-                                                </CmsBlockContextProvider>
-                                            </SnackbarProvider>
-                                        </DndProvider>
-                                    </MuiThemeProvider>
-                                </BlocksConfigProvider>
-                            </LocaleProvider>
-                        </IntlProvider>
-                    </SitesConfigProvider>
+                    <IntlProvider locale="en" messages={getMessages()}>
+                        <LocaleProvider resolveLocaleForScope={(scope: ContentScope) => scope.language}>
+                            <BlocksConfigProvider
+                                isBlockSupported={(block, scope) => {
+                                    if (scope.domain === "main") {
+                                        return true;
+                                    } else {
+                                        return (
+                                            block.name !== NewsDetailBlock.name &&
+                                            block.name !== NewsListBlock.name &&
+                                            block.name !== NewsLinkBlock.name
+                                        );
+                                    }
+                                }}
+                            >
+                                <MuiThemeProvider theme={theme}>
+                                    <DndProvider options={HTML5toTouch}>
+                                        <SnackbarProvider>
+                                            <CmsBlockContextProvider>
+                                                <ErrorDialogHandler />
+                                                <CurrentUserProvider>
+                                                    <RouterBrowserRouter>
+                                                        <GlobalStyle />
+                                                        <ContentScopeProvider>
+                                                            {({ match }) => (
+                                                                <Switch>
+                                                                    <Route
+                                                                        path={`${match.path}/preview`}
+                                                                        render={(props) => (
+                                                                            <SitePreview
+                                                                                resolvePath={(path: string, scope) => {
+                                                                                    return `/${scope.language}${path}`;
+                                                                                }}
+                                                                                {...props}
+                                                                            />
+                                                                        )}
+                                                                    />
+                                                                    <Route
+                                                                        render={() => (
+                                                                            <MasterLayout
+                                                                                headerComponent={MasterHeader}
+                                                                                menuComponent={AppMasterMenu}
+                                                                            >
+                                                                                <MasterMenuRoutes menu={masterMenuData} />
+                                                                            </MasterLayout>
+                                                                        )}
+                                                                    />
+                                                                </Switch>
+                                                            )}
+                                                        </ContentScopeProvider>
+                                                    </RouterBrowserRouter>
+                                                </CurrentUserProvider>
+                                            </CmsBlockContextProvider>
+                                        </SnackbarProvider>
+                                    </DndProvider>
+                                </MuiThemeProvider>
+                            </BlocksConfigProvider>
+                        </LocaleProvider>
+                    </IntlProvider>
                 </BuildInformationProvider>
             </ApolloProvider>
         </CometConfigProvider>
