+++ packages/admin/cms-admin/src/index.ts
@@ -191,10 +191,9 @@ export { useBlockPreview } from "./preview/block/useBlockPreview";
 export { openPreviewWindow, openSitePreviewWindow } from "./preview/openSitePreviewWindow";
 export { SitePreview } from "./preview/site/SitePreview";
 export { createRedirectsPage } from "./redirects/createRedirectsPage";
-export type { SiteConfig } from "./sitesConfig/SitesConfigContext";
-export { SitesConfigProvider } from "./sitesConfig/SitesConfigProvider";
-export { useSiteConfig } from "./sitesConfig/useSiteConfig";
-export { useSitesConfig } from "./sitesConfig/useSitesConfig";
+export type { SiteConfig } from "./siteConfigs/siteConfigsConfig";
+export { useSiteConfig } from "./siteConfigs/useSiteConfig";
+export { useSiteConfigs } from "./siteConfigs/useSiteConfigs";
 export { AzureAiTranslatorProvider } from "./translation/AzureAiTranslatorProvider";
 export { CurrentUserInterface, CurrentUserProvider, useCurrentUser, useUserPermissionCheck } from "./userPermissions/hooks/currentUser";
 export { UserPermissionsUserPageBasicDataPanel } from "./userPermissions/user/basicData/UserBasicData";
