+++ packages/admin/cms-admin/src/siteConfigs/siteConfigsConfig.ts
@@ -1,5 +1,3 @@
-import { createContext } from "react";
-
 import { type ContentScopeInterface } from "../contentScope/Provider";
 
 export interface SiteConfig {
@@ -9,9 +7,7 @@ export interface SiteConfig {
     sitePreviewApiUrl: string;
 }
 
-export interface SiteConfigApi<Configs = unknown> {
+export interface SiteConfigsConfig<Configs = unknown> {
     configs: Configs;
     resolveSiteConfigForScope: (configs: Configs, scope: ContentScopeInterface) => SiteConfig;
 }
-
-export const SiteConfigContext = createContext<SiteConfigApi | undefined>(undefined);
