--- packages/admin/cms-admin/src/sitesConfig/useSitesConfig.ts
@@ -1,15 +1,0 @@
-import { useContext } from "react";
-
-import { type SiteConfigApi, SiteConfigContext } from "./SitesConfigContext";
-
-export function useSitesConfig<Configs = unknown>(): SiteConfigApi<Configs> {
-    const context = useContext(SiteConfigContext);
-
-    if (!context) {
-        throw new Error(
-            "No SiteConfigContext instance can be found. Please ensure that you have called `SiteConfigProvider` higher up in your tree.",
-        );
-    }
-
-    return context as SiteConfigApi<Configs>;
-}
